[#275] Category useOutletContext 로 관리
- Context로 관리했던 데이터를 outletContext로 변경
