+++ front-end/src/components/common/TabBar/MainTabBar.tsx
@@ -14,10 +14,6 @@ interface TabBarInfo {
   path: string;
 }
 
-interface MainTabBarProps {
-  userId: number;
-}
-
 interface IconStyleProps {
   name: keyof typeof iconTypes;
   size?: string;
@@ -25,7 +21,7 @@ interface IconStyleProps {
   isActive?: boolean;
 }
 
-const MainTabBar = ({ userId }: MainTabBarProps) => {
+const MainTabBar = () => {
   const tabBarInfo: TabBarInfo[] = [
     { id: 1, icon: 'home', label: '홈화면', path: '/' },
     {
